colorize alert message, fix scripts for share
