+++ package.json
@@ -12,7 +12,8 @@
     "publish:dev": "npm run clean:vsix && node ./configs/publishDev",
     "build:release": "npm run lint && npm run clean && mkdir dist && webpack --progress --colors --config webpack.prod.config.js --output-path ./dist -p",
     "build:dev": "npm run lint && npm run clean && mkdir dist && webpack --progress --colors --config webpack.config.js --output-path ./dist -p",
-    "package:vsts:nobaseurl": "npm run build:release && node ./configs/packageDevVSTS",
+    "package:release:public": "npm run build:release && node ./configs/packageReleasePublic",
+    "package:release:share": "npm run build:release && node ./configs/packageReleaseForShare",
     "share:vsts": "npm run package:vsts:nobaseurl && node ./configs/sharevsts",
     "share:vsts:dev": "node ./configs/sharevstsDev",
     "dev": "webpack-dev-server --inline --progress --colors --port 3000",
